wil::resume_foreground moved to cppwinrt_helpers.h to avoid forward-declaration collisions
The forward declaration of the ::winrt::Microsoft namespace created
collisions with the ::Microsoft::WRL namespace when consumed from
inside the ::winrt namespace, because "Microsoft" resolves to
::winrt::Microsoft, but people who aren't using `winrt/Microsoft.*.h`
are expecting it to resolve to ::Microsoft.

We follow the `resource.h` pattern, where you include the headers
for the namespaces you care about, and then you include `cppwinrt_helpers.h`
to activate the helpers for those namespaces. You can include
`cppwinrt_headers.h` multiple times, and it will turn on the features
related to the namespaces you included since the last time you included
`cppwinrt_headers.h`.

This means that `cppwinrt_headers.h` does not introduce any namespaces.
It only takes advantage of pre-existing ones. If there is a namespace
name collision, it was introduced by the caller.
